pyproject.toml
[tool.poetry]
name = "hacksquad-discordpy"
version = "0.1.0"
description = ""
authors = ["Arpan Pandey <arpanpandey.aps@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.0.1"
python-dotenv = "^0.21.0"
aiohttp = "^3.8.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
